+++ src/asp/Sessions/StereoSessionGdal.h
@@ -149,7 +149,7 @@ namespace asp {
                       std::string const& camera_file,
                       std::string const& ba_prefix,
                       vw::Vector2 pixel_offset) const{
-      return load_adjusted_model(m_camera_loader.load_spot5_camera_model(camera_file),
+      return load_adjusted_model(m_camera_loader.load_spot15_camera_model(camera_file),
                                  image_file, camera_file, ba_prefix, pixel_offset);
     }
   };